+++ src/coord/mod.rs
@@ -29,7 +29,7 @@ mod numeric;
 mod ranged;
 
 #[cfg(feature = "chrono")]
-pub use datetime::{RangedDate, RangedDateTime};
+pub use datetime::{IntoMonthly, IntoYearly, RangedDate, RangedDateTime};
 pub use numeric::{
     RangedCoordf32, RangedCoordf64, RangedCoordi32, RangedCoordi64, RangedCoordu32, RangedCoordu64,
 };