+++ src/lib.rs
@@ -151,6 +151,16 @@ including bitmap, vector graph, piston window and WebAssembly.
     </div>
 </div>
 
+<div class="galleryItem">
+    <a href="https://plotters-rs.github.io/plotters-doc-data/slc-temp.png">
+        <img src="https://plotters-rs.github.io/plotters-doc-data/slc-temp.png" class="galleryItem"></img>
+    </a>
+    <div class="galleryText">
+        Monthly Time Coordinate
+        <a href="https://github.com/38/plotters/blob/master/examples/slc-temp.rs">[code]</a>
+    </div>
+</div>
+
 
 ## Table of Contents
   * [Gallery](#gallery)