+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Improvement
 
 - More examples are included
+- Date coordinate now support using monthly or yearly axis. This is usefuly when plotting some data in monthly or yearly basis.
 
 ## Plotters 0.2.5 (2019-09-07)